Split GenAI tasks, add Twitter posting, email templates, spam rejection
GenAI tasks extracted to separate files:
- Infrastructure/GenAiTasks/SpamFilterGenAiTask.cs
- Infrastructure/GenAiTasks/SeoAnalysisGenAiTask.cs
- Infrastructure/GenAiTasks/SocialMediaGenAiTask.cs

Social posting subscription:
- Set @refresh on future posts so subscription fires at publish time
- Move Social.GeneratedAt and DisableAutoPublish checks to subscription query
- Twitter posting via X API v2 with Bearer token from BlogConfig
- Reddit posting logged as pending RedditSharp 2.0 migration

Email templates:
- Switch from inline HTML to Scriban 7.2.0 templates
- Proper HTML email layout matching blog style (dark header #2c3e50)
- NewComment template: post link, comment body, author info, action buttons
- SpamDigest template: date summary, spam entries with preview, admin link

Spam rejection for repeat offenders:
- Commenter with >4 spam comments: return success but don't save comment
- User sees "posted soon" message but comment is silently discarded

BlogConfig: add TwitterBearerToken field + admin settings UI
