+++ RaccoonBlog.Web/Program.cs
@@ -20,6 +20,7 @@ using RaccoonBlog.Web.Services;
 using Raven.Client.Documents;
 using Raven.Client.Documents.Conventions;
 using Raven.Client.Documents.Operations.AI.Agents;
+using RaccoonBlog.Web.Infrastructure.GenAiTasks;
 using Raven.Client.Documents.Session;
 using Raven.Client.Http;
 using System;
@@ -352,272 +353,10 @@ static void ConfigureRefreshAndGenAiTasks(IDocumentStore store)
         log.Error(e, "Failed to configure refresh.");
     }
 
-    // All GenAI tasks expect an AI connection string named "ai-chat" to be configured
-    // in RavenDB Studio before starting the application. The connection string should point
-    // to a chat-capable model (e.g., OpenAI gpt-4o-mini or equivalent).
-
-    // GenAI spam filter task on PostComments collection
-    try
-    {
-        var config = new Raven.Client.Documents.Operations.AI.GenAiConfiguration
-        {
-            Name = "spam-filter",
-            Identifier = "spam-filter",
-            ConnectionStringName = "ai-chat",
-            Disabled = false,
-            Collection = "PostComments",
-            GenAiTransformation = new Raven.Client.Documents.Operations.AI.GenAiTransformation
-            {
-                Script = """
-                    const post = load(this.Post.Id);
-                    for(const comment of this.Comments) {
-                        if(comment.SpamCheckStatus !== 'Pending') 
-                            continue;
-
-                        ai.genContext({
-                            Post: {Title: post.Title, Tags: post.Tags},
-                            Id: comment.Id, 
-                            Author: comment.Author, 
-                            Body: comment.Body,
-                            Email: comment.Email,
-                            Url: comment.Url,
-                            UserHostAddress: comment.UserHostAddress, 
-                            UserAgent: comment.UserAgent,
-                            CommenterId: comment.CommenterId,
-                            PostCommentsId: id(this)
-                        });
-                    }
-                    """
-            },
-            Queries = [
-                new AiAgentToolQuery
-                {
-                    Name = "ReadPostComments",
-                    Description = """
-                                  Use this to read the _other_ comments on the same post to get more context. Returns all valid comments
-                                  on this post, so you can check is replying to another comment, instead of trying to evaluate it in isolation.
-                                  """,
-                    ParametersSampleObject = "{}",
-                    Query = """
-                            from PostComments where id() = $PostCommentsId
-                            """
-                }
-            ],
-            Prompt = """
-                You are a spam filter for a technical blog. Analyze this blog comment.
-                A spam comment typically includes irrelevant or promotional content,
-                excessive links, misleading information, or advertising intent.
-                A legitimate comment engages with the post, asks relevant questions,
-                or provides relevant feedback. You can see the title and tags of the post this
-                comment is in reply to.
-
-                Based on the comment content and metadata, determine if this comment is likely spam.
-                The expected language is English, if it is anything else, that is also a strong signal of spam.
-
-                If the comment may be spam, but can also be part of a conversation, you have the ReadPostComments
-                tool that you can invoke to get the other comments on the same post to get more context before making a decision.
-                """,
-            SampleObject = """
-                { "IsSpam": true, "Reason": "brief explanation why you think it is spam or not" }
-                """,
-            UpdateScript = """
-                const idx = this.Comments.findIndex(c => c.Id == $input.Id);
-                if (idx < 0) return;
-
-                if ($output.IsSpam) {
-                    var c = this.Comments[idx];
-                    c.IsSpam = true;
-                    c.SpamCheckStatus = 'Spam';
-                    this.Comments.splice(idx, 1);
-                    this.Spam.push(c);
-
-                    var post = load(this.Post.Id);
-                    if (post) {
-                        post.CommentsCount--;
-                        if (post.CommentsCount < 0) post.CommentsCount = 0;
-                    }
-
-                    var today = new Date().toISOString().split('T')[0];
-                    var digestId = 'SpamDigests/' + today;
-                    var tomorrow = new Date();
-                    tomorrow.setDate(tomorrow.getDate() + 1);
-                    tomorrow.setHours(8, 0, 0, 0);
-
-                    var spamEntry = {
-                        CommentId: $input.Id, 
-                        Author: $input.Author, 
-                        Body: $input.Body,
-                        PostId: this.Post.Id, 
-                        Timestamp: new Date().toISOString()
-                    };
-
-                    var dig = load(digestId) || {};
-                    dig.Type = 'SpamDigest';
-                    dig.View = 'SpamDigest';
-                    dig.DigestDate = today;
-                    dig.SpamComments = dig.SpamComments || [];
-                    dig.Count = dig.Count || 0;
-                    dig.SpamComments.push(spamEntry);
-                    dig.Count++;
-                    put(digestId, dig, {
-                        '@collection': 'EmailCommands',
-                        '@refresh': tomorrow.toISOString()
-                    });
-                } else {
-                    this.Comments[idx].SpamCheckStatus = 'Valid';
-
-                    if ($input.CommenterId) {
-                        var commenter = load($input.CommenterId);
-                        if (commenter !== null && !commenter.IsTrustedCommenter) {
-                            commenter.IsTrustedCommenter = true;
-                            put($input.CommenterId, commenter);
-                        }
-                    }
-
-                    var post = load(this.Post.Id);
-                    var postTitle = post ? post.Title : '';
-
-                    put('EmailCommands/new-comment-' + $input.Id, {
-                        Type: 'NewComment',
-                        View: 'NewComment',
-                        ReplyTo: $input.Email || '',
-                        Subject: 'Comment on: ' + postTitle + ' from ' + $input.Author,
-                        CommentId: $input.Id,
-                        Author: $input.Author || '',
-                        CommentBody: $input.Body || '',
-                        CommentEmail: $input.Email || '',
-                        CommentUrl: $input.Url || '',
-                        CreatedAt: new Date().toISOString(),
-                        IpAddress: $input.UserHostAddress || '',
-                        UserAgent: $input.UserAgent || '',
-                        CommenterId: $input.CommenterId || '',
-                        PostId: this.Post.Id || '',
-                        PostTitle: postTitle,
-                        PostSlug: post ? post.Slug : '',
-                        Key: post.ShowPostEvenIfPrivate
-                    }, {
-                        '@collection': 'EmailCommands'
-                    });
-                }
-                """
-        };
-        store.Maintenance.Send(new Raven.Client.Documents.Operations.AI.AddGenAiOperation(config));
-        log.Info("GenAI spam filter task created.");
-    }
-    catch (Exception e)
-    {
-        log.Error(e, "Failed to create GenAI spam filter task.");
-    }
-
-    // GenAI SEO analysis task on Posts collection
-    try
-    {
-        var config = new Raven.Client.Documents.Operations.AI.GenAiConfiguration
-        {
-            Name = "SEO Analysis",
-            Identifier = "seo-analysis",
-            ConnectionStringName = "ai-chat",
-            Disabled = false,
-            Collection = "Posts",
-            GenAiTransformation = new Raven.Client.Documents.Operations.AI.GenAiTransformation
-            {
-                Script = """
-                    ai.genContext({
-                        Title: this.Title,
-                        Body: this.Body,
-                        Tags: this.Tags
-                    });
-                    """
-            },
-            Prompt = """
-                You are an expert SEO analyst. Analyze the blog post provided and generate:
-
-                1. A compelling meta description (max 160 characters) that accurately summarizes the post and includes relevant keywords to improve search engine ranking. Write for humans, not search engines.
-
-                2. A list of 3-8 SEO keywords/keyphrases relevant to the post content. These should be terms people would search for to find this content. Include both short-tail and long-tail keywords where appropriate.
-
-                The post content is provided below. Analyze the title, body text, and existing tags.
-                """,
-            SampleObject = """
-                {
-                    "MetaDescription": "A concise, compelling meta description under 160 characters.",
-                    "Keywords": [ "primary keyword", "secondary keyword phrase", "related term" ]
-                }
-                """,
-            UpdateScript = """
-                this.SeoMetaDescription = $output.MetaDescription;
-                this.SeoKeywords = $output.Keywords;
-                this.SeoLastAnalyzedAt = new Date().toISOString();
-                """
-        };
-        store.Maintenance.Send(new Raven.Client.Documents.Operations.AI.AddGenAiOperation(config));
-        log.Info("GenAI SEO analysis task created.");
-    }
-    catch (Exception e)
-    {
-        log.Error(e, "Failed to create GenAI SEO analysis task.");
-    }
-
-    // GenAI social media text generation task on Posts collection
-    try
-    {
-        var config = new Raven.Client.Documents.Operations.AI.GenAiConfiguration
-        {
-            Name = "social-media",
-            Identifier = "social-media",
-            ConnectionStringName = "ai-chat",
-            Disabled = false,
-            Collection = "Posts",
-            GenAiTransformation = new Raven.Client.Documents.Operations.AI.GenAiTransformation
-            {
-                Script = """
-                    ai.genContext({
-                        Title: this.Title,
-                        Body: this.Body,
-                        Tags: this.Tags
-                    });
-                    """
-            },
-            Prompt = """
-                You are a social media manager for a technical blog about software development,
-                databases, and distributed systems. Generate engaging social media text for the
-                blog post provided.
-
-                1. Twitter: A concise, engaging tweet (max 250 characters, excluding URL which
-                   will be appended automatically). Should hook technical readers. Include 1-2
-                   relevant hashtags if natural. Do not include a URL.
-
-                2. Reddit: A compelling submission title for a programming subreddit audience.
-                   Should be informative and spark discussion. No clickbait. Max 300 characters.
-                """,
-            SampleObject = """
-                {
-                    "TwitterText": "Concise engaging tweet text with #relevantHashtag",
-                    "RedditTitle": "Compelling Reddit submission title for technical audience"
-                }
-                """,
-            UpdateScript = """
-                this.Social = this.Social || {};
-                this.Social.TwitterText = $output.TwitterText;
-                this.Social.RedditTitle = $output.RedditTitle;
-                this.Social.GeneratedAt = new Date().toISOString();
-
-                if (this.PublishAt) {
-                    var publishDate = new Date(this.PublishAt);
-                    if (publishDate > new Date()) {
-                        var metadata = getMetadata(this);
-                        metadata['@refresh'] = this.PublishAt;
-                    }
-                }
-                """
-        };
-        store.Maintenance.Send(new Raven.Client.Documents.Operations.AI.AddGenAiOperation(config, StartingPointChangeVector.LastDocument));
-        log.Info("GenAI social media task created.");
-    }
-    catch (Exception e)
-    {
-        log.Error(e, "Failed to create GenAI social media task.");
-    }
+    // Register GenAI tasks (each in its own file under Infrastructure/GenAiTasks/)
+    SpamFilterGenAiTask.Register(store);
+    SeoAnalysisGenAiTask.Register(store);
+    SocialMediaGenAiTask.Register(store);
 
     // Start subscription workers
     RaccoonBlog.Web.Infrastructure.EmailSubscription.Start(store);
